implement new way to register arbitrary skills
now that i'm moving towards the workflow side of the project,
it's time to make decisions. this was kinda sneak-peeked
during the committing of 4051445b but i wasn't confident
if this was the right way to move forward.

now invigorated with confidence, i decided that i should maintain
two ways for the project to be integrated similarly to how React
supported functional and class-based components at one time.

however, instead of maintaining these two methods as entirely
separate entities, i made it so the functional method of using QGF
instead converges into the class-based method, albeit registration
isn't done automatically as a result.

exposing a builder-based approach prevents the accidental alienation
of the target audience, that being roblox developers of all sorts. most
roblox developers use luau, not roblox-ts, and composing a class
structure in luau in a way that roblox-ts can consume is not something
one should expect from a roblox developer as such. therefore, exposing
something that can instead be accessed and called upon instead of
manifesting something and then appending to it is a much more preferable
approach.

for months i've been trying to think of a way to have characters and
skills cleanly registered into the qgf system, and my brain kept
pointing to decorators but in the early stages we had only builders, and
you couldn't decorate const variables then. however, a year later with
myself being a much better programmer, i was able to integrate after a
long internal battle. it was worth it in the end though.

Signed-off-by: nowaaru@proton.me <vert@leanbox.>